requirements.txt
django
bleach
django-countries
django-debug-toolbar
django-djconfig
django-haystack
django-infinite-scroll-pagination
django-messages
mistune
MySQL-python
Pillow
South
wheel
Whoosh
django-resized
Faker
pytz
django-autoslug
django-pipeline
django-newsletter
sorl-thumbnail
django-extensions
python-card-me
ldif3
chardet
surlex
six
unicodecsv
django-oauth-toolkit
django-cors-headers
requests
django-guardian
uni-slugify